More moves and deletes
