+++ fairlearn/reductions/__init__.py
@@ -1,5 +1,4 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License.
 
-from fairlearn.reductions.grid_search_classification import GridSearchClassification  # noqa: F401
-from fairlearn.reductions.grid_search.regression import GridSearchRegression  # noqa: F401
+from fairlearn.reductions.grid_search.grid_search import GridSearch  # noqa: F401
