+++ fairlearn/reductions/grid_search/grid_search.py
@@ -5,10 +5,9 @@ import copy
 import numpy as np
 
 from fairlearn.metrics import DemographicParity, BoundedGroupLoss
-from fairlearn.reductions.grid_search import utilities as utilities
 
 
-class GridSearchClassification:
+class GridSearch:
     def __init__(self,
                  learner,
                  fairness_metric,
@@ -24,18 +23,20 @@ class GridSearchClassification:
 
     def fit(self, X, Y, protected_attribute, lagrange_multipliers=None):
         if isinstance(self.fairness_metric, DemographicParity):
-            return self.fit_classification(X, Y, protected_attribute, lagrange_multipliers)
+            return self._fit_classification(X, Y, protected_attribute, lagrange_multipliers)
+        elif isinstance(self.fairness_metric, BoundedGroupLoss):
+            return self._fit_regression(X, Y, protected_attribute, lagrange_multipliers)
         else:
             raise RuntimeError("Can't get here")
 
-    def fit_classification(self, X, Y, protected_attribute, lagrange_multipliers=None):
+    def _fit_classification(self, X, Y, protected_attribute, lagrange_multipliers=None):
         # Verify we have a binary classification problem
         unique_labels = np.unique(Y)
         if not set(unique_labels).issubset({0, 1}):
             raise RuntimeError("Supplied Y labels are not 0 or 1")
 
         # Extract required statistics from protected_attribute
-        p0, p1, a0_val = utilities.generate_protected_attribute_info(protected_attribute)
+        p0, p1, a0_val = self._generate_protected_attribute_info(protected_attribute)
 
         # If not supplied, generate array of trial lagrange multipliers
         if lagrange_multipliers is None:
@@ -73,6 +74,29 @@ class GridSearchClassification:
         # Designate a 'best' model
         self.best_model = self._select_best_model(X, Y, protected_attribute, self.all_models)
 
+    def _fit_regression(self, X, Y, protected_attribute, tradeoffs=None):
+        # Extract required statistics from protected_attribute
+        p0, p1, a0_val = self._generate_protected_attribute_info(protected_attribute)
+
+        if tradeoffs is None:
+            tradeoffs = np.linspace(0, 1, self.number_of_lagrange_multipliers)
+
+        self.all_models = []
+        for tradeoff in tradeoffs:
+            weight_func = np.vectorize(self._regression_weight_function)
+            weights = weight_func(protected_attribute,
+                                  tradeoff,
+                                  p0, p1, a0_val)
+
+            current_learner = copy.deepcopy(self.learner)
+            current_learner.fit(X, Y, sample_weight=weights)
+
+            self.all_models.append({"model": current_learner,
+                                    "tradeoff": tradeoff})
+
+        # Designate a 'best' model
+        self.best_model = self._select_best_model(X, Y, protected_attribute, self.all_models)
+
     def predict(self, X):
         return self.best_model["model"].predict(X)
 
@@ -91,6 +115,18 @@ class GridSearchClassification:
         else:
             return 2 * y_val - 1 + L
 
+    def _generate_protected_attribute_info(self, protected_attribute):
+        unique_labels, counts = np.unique(
+            protected_attribute, return_counts=True)
+        if len(unique_labels) > 2:
+            raise RuntimeError("Protected Attribute contains "
+                               "more than two unique values")
+
+        p0 = counts[0] / len(protected_attribute)
+        p1 = 1 - p0
+
+        return p0, p1, unique_labels[0]
+
     def _generate_weights(self, y, protected_attribute, L, p_ratio, a0_val):
         weight_func = np.vectorize(self._weight_function)
         return weight_func(y, protected_attribute, L, p_ratio, a0_val)
@@ -103,3 +139,9 @@ class GridSearchClassification:
 
         best_model = max(model_list, key=lambda x: x["quality"])
         return best_model
+
+    def _regression_weight_function(self, a_val, trade_off, p0, p1, a0_val):
+        if a_val == a0_val:
+            return trade_off / p0
+        else:
+            return (1 - trade_off) / p1
